backend/foodgram/requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==2.2.16
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
netaddr==0.8.0
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pylibscrypt==2.0.0
python3-openid==3.2.0
pytz==2020.1
pywin32==303
requests==2.26.0
requests-oauthlib==1.3.1
scrypt==0.8.20
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.3.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.9
varint==1.0.2
voluptuous==0.13.0
wcwidth==0.2.5
zipp==3.7.0
